==> notebooks/1.0-ry-extract-titanic-data.ipynb <==
!C:\Users\yrahu\Anaconda3\python -m pip install python-dotenv

from dotenv import load_dotenv, find_dotenv

dotenv_path = find_dotenv()
load_dotenv(dotenv_path)

import os
KAGGLE_USERNAME = os.environ.get("KAGGLE_USERNAME")


import requests
from requests import session

payload = {
    "action": "login",
  	"username": os.environ.get("KAGGLE_USERNAME"),
  	"password": os.environ.get("KAGGLE_PASSWORD")
}

url = "https://www.kaggle.com/c/titanic/download/train.csv"

with session() as c:
    c.post("https://www.kaggle.com/account/login", data=payload)
    response = c.get(url)
    print(response.text)

def extract_data(url, file_path):
    """
    Extract data from Kaggle
    """
    with session() as c:
        c.post("https://www.kaggle.com/account/login", data=payload)
        with open(file_path, "wb") as f:
            response = c.get(url, stream=True)
            for block in response.iter_content(1024):
                f.write(block)

train_url = "https://www.kaggle.com/c/titanic/download/train.csv"
test_url = "https://www.kaggle.com/c/titanic/download/test.csv"

raw_data_path = os.path.join(os.path.pardir, "data", "raw")
train_data_path = os.path.join(raw_data_path, "train.csv")
test_data_path = os.path.join(raw_data_path, "test.csv")

extract_data(train_url, train_data_path)
extract_data(test_url, test_data_path)

!ls -l ../data/raw

get_raw_data_script_file = os.path.join(os.path.pardir, "src", "data", "get_raw_data.py")

%%writefile $get_raw_data_script_file
# -*- coding: utf-8 -*-
from dotenv import load_dotenv, find_dotenv
import os
import requests
from requests import session
import logging

payload = {
    "action": "login",
    "username": os.environ.get("KAGGLE_USERNAME"),
    "password": os.environ.get("KAGGLE_PASSWORD")
}

def extract_data(url, file_path):
    """
    Extract data from Kaggle
    """
    with session() as c:
        c.post("https://www.kaggle.com/account/login", data=payload)
        with open(file_path, "wb") as f:
            response = c.get(url, stream=True)
            for block in response.iter_content(1024):
                f.write(block)


def main(project_dir):
    logger = logging.getLogger(__name__)
    logger.info("getting raw data")

    dotenv_path = find_dotenv()
    load_dotenv(dotenv_path)

    train_url = "https://www.kaggle.com/c/titanic/download/train.csv"
    test_url = "https://www.kaggle.com/c/titanic/download/test.csv"

    raw_data_path = os.path.join(project_dir, "data", "raw")
    train_data_path = os.path.join(raw_data_path, "train.csv")
    test_data_path = os.path.join(raw_data_path, "test.csv")

    extract_data(train_url, train_data_path)
    extract_data(test_url, test_data_path)
    logger.info("done!")


if __name__ == "__main__":
    project_dir = os.path.join(os.path.dirname(__file__), os.pardir, os.pardir)
    
    log_fmt = "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
    logging.basicConfig(level=logging.INFO, format=log_fmt)
    
    dotenv_path = find_dotenv()
    load_dotenv(dotenv_path)
    
    main(project_dir)

!C:\Users\yrahu\Anaconda3\python $get_raw_data_script_file
