==> notebooks/3.0-ry-building-predictive-model.ipynb <==
# Building Predictive Model

import pandas as pd
import numpy as np
import os

## Import Data

processed_data_path = os.path.join(os.path.pardir, "data", "processed")
train_file_path = os.path.join(processed_data_path, "train.csv")
test_file_path = os.path.join(processed_data_path, "test.csv")

train_df = pd.read_csv(train_file_path, index_col="PassengerId")
test_df = pd.read_csv(test_file_path, index_col="PassengerId")

train_df.info()

test_df.info()

## Data Preparation

X = train_df.loc[:,:"AgeState_Child"].as_matrix().astype("float")
Y = train_df["Survived"].ravel()  # Creates flattened 1D array

print(X.shape, Y.shape)

from sklearn.model_selection import train_test_split  # sklearn.model_selection from sklearn 0.19+
X_train, X_test, y_train, y_test = train_test_split(X, Y, test_size=0.2, random_state=0)
print(X_train.shape, X_test.shape)
print(y_train.shape, y_test.shape)

import sklearn

sklearn.__version__

## Building the baseline model

from sklearn.dummy import DummyClassifier

dummy_model = DummyClassifier(strategy="most_frequent", random_state=0)

dummy_model.fit(X_train, y_train)

print("Score for baseline model = {0:.2f}".format(dummy_model.score(X_test, y_test)))

# Performance metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

print("Accuracy for baseline model = {0:.2f}".format(accuracy_score(y_test, dummy_model.predict(X_test))))

## First Kaggle submission

test_X = test_df.as_matrix().astype("float")

predictions = dummy_model.predict(test_X)

df_submission = pd.DataFrame({"PassengerId": test_df.index, "Survived": predictions})

df_submission.head()

submission_data_path = os.path.join(os.path.pardir, "data", "external")
submission_file_path = os.path.join(submission_data_path, "01_dummy.csv")

df_submission.to_csv(submission_file_path, index=False)

def get_submission_file(model, filename):
    test_X = test_df.as_matrix().astype("float")
    predictions = model.predict(test_X)
    
    df_submission = pd.DataFrame({"PassengerId": test_df.index, "Survived": predictions})
    
    submission_data_path = os.path.join(os.path.pardir, "data", "external")
    submission_file_path = os.path.join(submission_data_path, filename)
    df_submission.to_csv(submission_file_path, index=False)

## Logistic regression model

from sklearn.linear_model import LogisticRegression

model_lr_1 = LogisticRegression(random_state=0)

model_lr_1.fit(X_train, y_train)

print("Score for LR model: {0:.2f}".format(model_lr_1.score(X_test, y_test)))

predictions = model_lr_1.predict(X_test)
print("Accuracy score: {0:.2f}".format(accuracy_score(y_test, predictions)))
print("Confusion matrix:\n{0}".format(confusion_matrix(y_test, predictions)))
print("Precision score: {0:.2f}".format(precision_score(y_test, predictions)))
print("Recall score: {0:.2f}".format(recall_score(y_test, predictions)))

model_lr_1.coef_

## Second Kaggle Submission

get_submission_file(model_lr_1, "02_lr.csv")

## Hyperparameter Optimization

model_lr = LogisticRegression(random_state=0)

from sklearn.model_selection import GridSearchCV

parameters = {"C": [1.0, 10.0, 50.0, 100.0, 1000.0], "penalty": ["l1", "l2"]}
clf = GridSearchCV(model_lr, param_grid=parameters, cv=3)

clf.fit(X_train, y_train)

clf.best_params_

clf.best_score_

print("Score for logistic regression model: {0:.2f}".format(clf.score(X_test, y_test)))

## Third Kaggle Submission

get_submission_file(clf, "03_lr.csv")

## Feature Normalization and Standardization

from sklearn.preprocessing import MinMaxScaler, StandardScaler

### Feature Normalization

scaler = MinMaxScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_train_scaled[:,0].min(), X_train_scaled[:,0].max()

# Normalize test data
X_test_scaled = scaler.fit_transform(X_test)

### Feature Standardization

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

### Create Model after Standardization

model_lr = LogisticRegression()
parameters = {"C": [0.5, 1.0, 2.0, 5.0, 10.0, 50.0, 100.0], "penalty": ["l1", "l2"]}
clf = GridSearchCV(model_lr, param_grid=parameters, cv=3)
clf.fit(X_train_scaled, y_train)

clf.best_params_

clf.best_score_

print("Score for standardized model: {0:.2f}".format(clf.score(X_test_scaled, y_test)))

## Model Persistence

import pickle

model_file_path = os.path.join(os.path.pardir, "models", "lr_model.pkl")
scaler_file_path = os.path.join(os.path.pardir, "models", "lr_scaler.pkl")

model_file_pickle = open(model_file_path, "wb")
scaler_file_pickle = open(scaler_file_path, "wb")

pickle.dump(clf, model_file_pickle)
pickle.dump(scaler, scaler_file_pickle)

model_file_pickle.close()
scaler_file_pickle.close()

### Load the persisted file

model_file_pickle = open(model_file_path, "rb")
scaler_file_pickle = open(scaler_file_path, "rb")

clf_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)

model_file_pickle.close()
scaler_file_pickle.close()

clf_loaded
