==> notebooks/4.0-ry-building-machine-learning-api.ipynb <==
## Hello World using Flask

import os
script_file = os.path.join(os.path.pardir, "src", "models", "hello_world_api.py")

%%writefile $script_file
from flask import Flask, request

app = Flask(__name__)

@app.route("/api", methods=["POST"])
def say_hello():
    data = request.get_json(force=True)
    name = data["name"]
    return "Hello, {0}".format(name)

if __name__ == "__main__":
    app.run(port=9001, debug=True)

import json
import requests

url = "http://localhost:9001/api"
data = json.dumps({"name": "Rahul"})
r = requests.post(url, data)

print(r.text)

## ML API

ml_api_file = os.path.join(os.path.pardir, "src", "models", "ml_api.py")

%%writefile $ml_api_file
from flask import Flask, request
import pandas as pd
import numpy as np
import pickle
import json
import os

app = Flask(__name__)

# Load model and scaler
model_path = os.path.join(os.path.pardir, os.path.pardir, "models")
model_file = os.path.join(model_path, "lr_model.pkl")
scaler_file = os.path.join(model_path, "lr_scaler.pkl")

scaler = pickle.load(open(scaler_file, "rb"))
model = pickle.load(open(model_file, "rb"))

columns = ["Age", "Fare", "FamilySize",
       "IsMother", "IsMale", "Deck_A", "Deck_B", "Deck_C", "Deck_D", 
       "Deck_E", "Deck_F", "Deck_G", "Deck_Z", "Pclass_1", "Pclass_2", 
       "Pclass_3", "Title_Lady", "Title_Master", "Title_Miss", "Title_Mr", 
       "Title_Mrs", "Title_Officer", "Title_Sir", "Fare_Bin_Very_Low", 
       "Fare_Bin_Low", "Fare_Bin_High", "Fare_Bin_Very_High", "Embarked_C", 
       "Embarked_Q", "Embarked_S", "AgeState_Adult", "AgeState_Child"]

@app.route("/api", methods=["POST"])
def make_prediction():
    # Read JSON from the request, convert to JSON string
    data = json.dumps(request.get_json(force=True))
    
    # Create Pandas DataFrame
    df = pd.read_json(data)
    
    # Extract PassengerId
    pids = df["PassengerId"].ravel()
    
    X = df[columns].as_matrix().astype("float")
    X_scaled = scaler.transform(X)
    
    predictions = model.predict(X_scaled)
    df_response = pd.DataFrame({"PassengerId": pids, "Predicted": predictions})
    
    return df_response.to_json()

if __name__ == "__main__":
    app.run(port=9001, debug=True)

### Invoking ML API

import os
import pandas as pd

processed_data_path = os.path.join(os.path.pardir, "data", "processed")
train_file_path = os.path.join(processed_data_path, "train.csv")
train_df = pd.read_csv(train_file_path)

survived_passengers = train_df[train_df.Survived == 1][:5]

survived_passengers

import requests

def make_api_req(data):
    url = "http://localhost:9001/api"
    r = requests.post(url, data)
    return r.json()

make_api_req(survived_passengers.to_json())

result = make_api_req(train_df.to_json())
df_result = pd.read_json(json.dumps(result))
df_result.head()
